Remove scipy from the requirements
The requirement was added with an upper bound on October 27 2017
when scipy v1.0.0 was released, but it was broken. Therefore we had
to fix the upper bound but now scipy v1.0.1 is out and we can remove
it. Especially since having it as a requirement was failing the builds
on readthedocs due to it running out of memory